layout/base/tests/bug1080361_inner.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1080361
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1080361</title>
  <meta name="author" content="Maksim Lebedev" />
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <style>
    #target, #mediator, #listener { background: yellow; margin: 10px; }
  </style>
  <script type="application/javascript">
    var target = undefined;
    var mediator = undefined;
    var listener = undefined;
    var test_target_down = false;
    var test_target_up = false;
    var test_first_exc = false;
    var test_second_exc = false;
    var test_third_exc = false;
    var test_fourth_exc = false;
    var test_listener = false;

    function TargetDownHandler(event) {
      logger("Target receive event: " + event.type);
      test_target_down = true;
      try {
        logger("target.setPointerCapture()");
        target.setPointerCapture(31415);
      } catch(exc) {
        test_first_exc = true;
        parent.is(exc.name, "InvalidPointerId", "Exception InvalidPointerId should be fired");
      }
      try {
        logger("mediator.setPointerCapture()");
        mediator.remove();
        mediator.setPointerCapture(event.pointerId);
      } catch(exc) {
        test_second_exc = true;
        parent.is(exc.name, "InvalidStateError", "Exception InvalidStateError should be fired");
      }
      try {
        logger("listener.setPointerCapture()");
        listener.setPointerCapture(event.pointerId);
      } catch(exc) {
        test_third_exc = true;
      }
    }
    function TargetUpHandler(event) {
      logger("Target receive event: " + event.type);
      test_target_up = true;
      try {
        logger("target.setPointerCapture()");
        target.setPointerCapture(event.pointerId);
      } catch(exc) {
        test_fourth_exc = true;
      }
    }
    function ListenerHandler(event) {
      logger("Listener receive event: " + event.type);
      test_listener = true;
      listener.releasePointerCapture(event.pointerId);
    }
    function logger(message) {
      console.log(message);
      var log = document.getElementById('log');
      log.innerHTML = message + "<br>" + log.innerHTML;
    }

    function prepareTest() {
      parent.turnOnPointerEvents(executeTest);
    }
    function executeTest()
    {
      logger("executeTest");
      target = document.getElementById("target");
      mediator = document.getElementById("mediator");
      listener = document.getElementById("listener");
      target.addEventListener("pointerdown",          TargetDownHandler);
      target.addEventListener("pointerup",            TargetUpHandler);
      listener.addEventListener("gotpointercapture",  ListenerHandler);
      var rect = target.getBoundingClientRect();
      synthesizeMouse(target, rect.width/2, rect.height/2, {type: "mousedown"});
      synthesizeMouse(target, rect.width/2, rect.height/2, {type: "mousemove"});
      synthesizeMouse(target, rect.width/2, rect.height/2, {type: "mouseup"});
      finishTest();
    }
    function finishTest() {
      setTimeout(function() {
        parent.is(test_target_down, true,   "pointerdown event should be received by target");
        parent.is(test_target_up,   true,   "pointerup event should be received by target");
        parent.is(test_first_exc,   true,   "first exception should be thrown");
        parent.is(test_second_exc,  true,   "second exception should be thrown");
        parent.is(test_third_exc,   false,  "third exception should not be thrown");
        parent.is(test_fourth_exc,  false,  "fourth exception should not be thrown");
        parent.is(test_listener,    true,   "listener should receive gotpointercapture event");
        logger("finishTest");
        parent.finishTest();
      }, 1000);
    }
  </script>
</head>
<body onload="prepareTest()">
  <a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1080361">Mozilla Bug 1080361</a>
  <div id="target">div id=target</div>
  <div id="mediator">div id=mediator</div>
  <div id="listener">div id=listener</div>
  <pre id="log"></pre>
</body>
</html>
